+++ src/main/kotlin/io/runescript/plugin/ide/config/RsConfig.kt
@@ -10,6 +10,8 @@ object RsConfig {
         "varp",
         "varbit",
         "varc",
+        "varclan",
+        "varclansetting",
         "null",
         "int",
         "string",