+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Import functionality for Neptune modules.
 - Local/system settings to manage launcher JRE and Neptune home.
 - Icons for script declarations.
+- Support for varclan and varclansetting.
 
 ### Changed